refactor : setter method 제거
- synchronized 적용
