--- src/main/java/techcourse/myblog/HelloWorldController.java
@@ -1,8 +1,0 @@
-package techcourse.myblog;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class HelloWorldController {
-
-}
